moved create flow in side panel
