+++ src/bp/ui-studio/src/web/views/FlowBuilder/toolbar.jsx
@@ -49,26 +49,6 @@ class Toolbar extends React.Component {
       })
     }
 
-    const promptNewFlow = () => {
-      let name = prompt('Enter the name of the new flow')
-
-      if (!name) {
-        return
-      }
-
-      name = name.replace(/\.flow\.json$/i, '')
-
-      if (/[^A-Z0-9-_\/]/i.test(name)) {
-        return alert('ERROR: The flow name can only contain letters, numbers, underscores and hyphens.')
-      }
-
-      if (_.includes(this.props.flowsNames, name + '.flow.json')) {
-        return alert('ERROR: This flow already exists')
-      }
-
-      this.props.onCreateFlow(name)
-    }
-
     const noSkills = (
       <MenuItem eventKey="1" disabled={true}>
         No skills installed
@@ -92,14 +72,6 @@ class Toolbar extends React.Component {
     return (
       <div className={style.wrapper}>
         <div className={style.toolbar}>
-          <Button className={style.btn} bsStyle="default" onClick={promptNewFlow}>
-            <OverlayTrigger placement="bottom" overlay={createTooltip('addFlow', 'Create new flow')}>
-              <i className="material-icons">create_new_folder</i>
-            </OverlayTrigger>
-          </Button>
-
-          <div className={style.separator} />
-
           <Button
             className={style.btn}
             bsStyle="default"
@@ -110,8 +82,6 @@ class Toolbar extends React.Component {
               <i className="material-icons">save</i>
             </OverlayTrigger>
           </Button>
-
-          <div className={style.separator} />
 
           <Button className={style.btn} bsStyle="default" disabled={!this.props.canUndo} onClick={this.props.undo}>
             <OverlayTrigger placement="bottom" overlay={createTooltip('undo', 'Undo')}>
